Abstract out of a class to reduce duplicated logic between InMemoryIdentityService and PersistentIdentityService
